backend/requirements.txt
Django==3.2.16
django-filter==22.1
djangorestframework==3.12.4
drf-base64==2.0
gunicorn==20.0.4
psycopg2-binary==2.9.6
psycopg2==2.9.6
python-dotenv==1.0.0
pytest==6.2.4
pytest-django==4.4.0
PyJWT==2.7.0
flake8==6.0.0
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
djoser==2.2.0
Pillow==9.5.0
reportlab==4.0.4
